feat: migrate shaders to assets, optimize temporal math, and enable profiling
- Synchronized GPU shader layouts with alignas(16) C++ structs
- Moved shaders to app/src/main/assets/shaders and implemented AssetManager loading
- Optimized temporal math using integer epoch seconds and GPU-side decomposition
- Added high-resolution ATrace markers for Perfetto orchestration analysis
- Configured <profileable> in AndroidManifest and enabled MEASUREMENT in release build

+++ app/src/main/cpp/Components/TimestampComponent.h
@@ -1,11 +1,13 @@
 #ifndef ZEROSTALLCAM_COMPONENTS_TIMESTAMPCOMPONENT_H
 #define ZEROSTALLCAM_COMPONENTS_TIMESTAMPCOMPONENT_H
 
+#include <cstdint>
+
 namespace ecs {
 
 struct alignas(16) TimestampComponent {
-    float epochSeconds;
-    float p1, p2, p3; // Padding to 16 bytes
+    int32_t epochSeconds;
+    int32_t p1, p2, p3; // Padding to 16 bytes
 };
 
 } // namespace ecs
